refactor(shared-data): module python bindings
Add bindings to load module definitions easily and, more importantly,
rich types for module definition access and tests for same.

This shows how we'll use the bindings here: to provide the kind of
little functions that used to be scattered around the opentrons module,
and to provide rich types using TypedDict for the json definition files,
which will allow much safer lookup of data from the dicts.

To make sure that the type definitions actually match what's in the
types, we check them using the python package typeguard[1], which is
like mypy but invokable at run time much more easily. We don't want to
do this in the actual shipping package, but this is a perfect thing to
use in the now-existing tests to iterate through the existing definition
files (which are the sources of truth about the structure of the types,
since they are json schema checked) and make sure the types match them.

[1]: https://typeguard.readthedocs.io/en/latest/