refactor(api): use new module def accessors
This actually uses the module definition accessors in the protocol api!
Note some changes required by the presence of actual types; in
particular, we can't quite discriminate between the two module def
versions based on the presence or absence of a key (tragically) so
there's a couple type ignores in there.